+++ parquet-column/src/main/java/org/apache/parquet/column/statistics/geospatial/GeospatialTypes.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.parquet.column.statistics.geometry;
+package org.apache.parquet.column.statistics.geospatial;
 
 import java.util.HashSet;
 import java.util.Set;
