+++ parquet-column/src/test/java/org/apache/parquet/column/statistics/geospatial/TestBoundingBox.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.parquet.column.statistics.geometry;
+package org.apache.parquet.column.statistics.geospatial;
 
 import org.junit.Assert;
 import org.junit.Test;
